JWS is now fully abstract, JWE converson function, code cleanup

+++ JWT-Library/Lib/Helpers/NumericDate.cs
@@ -15,20 +15,20 @@ namespace JWTLib
         /// Gets todays date in numeric value (seconds since epoch)
         /// </summary>
         /// <returns></returns>
-        public static double Today()
+        public static long Today()
         {
             // Calculate and return the number of seconds from epoch 
-            return Math.Floor((DateTime.Now - DateTimeOffset.UnixEpoch).TotalSeconds);
+            return (long)Math.Floor((DateTime.Now - DateTimeOffset.UnixEpoch).TotalSeconds);
         }
 
         /// <summary>
         /// Converts the given date to a numeric date (seconds since epoch)
         /// </summary>
         /// <returns></returns>
-        public static double Convert(DateTime date)
+        public static long Convert(DateTime date)
         {
             // Calculate and return the number of seconds from epoch since the given date 
-            return Math.Floor((date - DateTimeOffset.UnixEpoch).TotalSeconds);
+            return (long)Math.Floor((date - DateTimeOffset.UnixEpoch).TotalSeconds);
         }
     }
 }
